Remove dead code and move Dialogue event system under Di
