--- Fantasy-Quest/Assets/Prefabs/ScenesManagers/Gameplay/DialogueEventSystem.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bf8012af5586bfd45a340c0b4c5d833c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
