+++ Fantasy-Quest/Assets/Scenes/Prod/House.unity
@@ -2606,8 +2606,8 @@ MonoBehaviour:
   m_ShadowVolumeIntensityEnabled: 1
   m_ShadowVolumeIntensity: 0.869
   m_LocalBounds:
-    m_Center: {x: -1.6427499, y: -5.62475, z: 0}
-    m_Extent: {x: 2.1059499, y: 1.83265, z: 0}
+    m_Center: {x: -1.6475501, y: -5.6289997, z: 0}
+    m_Extent: {x: 2.10815, y: 1.8309, z: 0}
   m_PointLightInnerAngle: 359.6
   m_PointLightOuterAngle: 359.6
   m_PointLightInnerRadius: 2
@@ -3734,8 +3734,8 @@ MonoBehaviour:
   m_ShadowVolumeIntensityEnabled: 1
   m_ShadowVolumeIntensity: 0.869
   m_LocalBounds:
-    m_Center: {x: -1.34225, y: -5.5739, z: 0}
-    m_Extent: {x: 1.75645, y: 1.9265001, z: 0}
+    m_Center: {x: -1.3475001, y: -5.5634, z: 0}
+    m_Extent: {x: 1.7602, y: 1.92, z: 0}
   m_PointLightInnerAngle: 359.6
   m_PointLightOuterAngle: 359.6
   m_PointLightInnerRadius: 2
@@ -9054,8 +9054,8 @@ MonoBehaviour:
   m_ShadowVolumeIntensityEnabled: 1
   m_ShadowVolumeIntensity: 0.869
   m_LocalBounds:
-    m_Center: {x: 0.46254998, y: -5.5828, z: 0}
-    m_Extent: {x: 1.45665, y: 1.961, z: 0}
+    m_Center: {x: 0.46749997, y: -5.5858502, z: 0}
+    m_Extent: {x: 1.4553, y: 1.96235, z: 0}
   m_PointLightInnerAngle: 359.6
   m_PointLightOuterAngle: 359.6
   m_PointLightInnerRadius: 2
@@ -15769,8 +15769,8 @@ MonoBehaviour:
   m_ShadowVolumeIntensityEnabled: 1
   m_ShadowVolumeIntensity: 0.869
   m_LocalBounds:
-    m_Center: {x: -2.8959498, y: -7.51785, z: 0}
-    m_Extent: {x: 1.9810499, y: 1.2692499, z: 0}
+    m_Center: {x: -2.8917, y: -7.5184, z: 0}
+    m_Extent: {x: 1.9872, y: 1.2684002, z: 0}
   m_PointLightInnerAngle: 359.6
   m_PointLightOuterAngle: 359.6
   m_PointLightInnerRadius: 2
@@ -17008,8 +17008,8 @@ MonoBehaviour:
   m_ShadowVolumeIntensityEnabled: 1
   m_ShadowVolumeIntensity: 0.869
   m_LocalBounds:
-    m_Center: {x: -0.6571, y: -7.4294004, z: 0}
-    m_Extent: {x: 2.6107001, y: 1.3264, z: 0}
+    m_Center: {x: -0.65715003, y: -7.4392, z: 0}
+    m_Extent: {x: 2.60045, y: 1.3225999, z: 0}
   m_PointLightInnerAngle: 359.6
   m_PointLightOuterAngle: 359.6
   m_PointLightInnerRadius: 2
@@ -17870,63 +17870,6 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 0a96211b37e72734996bb17dda89583a, type: 3}
---- !u!1001 &6690177770861202161
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    serializedVersion: 3
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 7949403621630559023, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_Name
-      value: DialogueEventSystem
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 9075148890220742808, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-    m_RemovedGameObjects: []
-    m_AddedGameObjects: []
-    m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: bf8012af5586bfd45a340c0b4c5d833c, type: 3}
 --- !u!114 &6732909966977288568 stripped
 MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 448578416113120038, guid: 28d9e3495a29edc46b1365288dc22921, type: 3}
@@ -18650,7 +18593,6 @@ SceneRoots:
   - {fileID: 7519323075250086346}
   - {fileID: 6446133136072648679}
   - {fileID: 6792728119773763104}
-  - {fileID: 6690177770861202161}
   - {fileID: 8855840689283057129}
   - {fileID: 4014933508526661021}
   - {fileID: 1801996417}
