--- Fantasy-Quest/Assets/Scripts/Common/ConditionsGate.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9fe92628b3a84f9408777750c2962e79
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
