--- Fantasy-Quest/Assets/Scripts/Common/FloatRange.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using Sirenix.OdinInspector;
-
-namespace Common
-{
-    [Serializable]
-    public struct FloatRange
-    {
-        [MaxValue(nameof(Max))]
-        public float Min;
-
-        [MinValue(nameof(Min))]
-        public float Max;
-
-        public float GetRandomFloatInRange()
-        {
-            return UnityEngine.Random.Range(Min, Max);
-        }
-    }
-}
