--- Fantasy-Quest/Assets/Scripts/Common/FloatRange.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 607e2a0be71bed041b17725fe22846f9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
