--- Fantasy-Quest/Assets/Scripts/Common/ISceneSingleton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: faf1a24b7dfac524bb6ddd86c7c71bad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
