--- Fantasy-Quest/Assets/Scripts/Common/MathfTools.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2ca04f848c864864ea28fd81a5a0b05f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
