--- Fantasy-Quest/Assets/Scripts/Common/Result.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b9386c5715a505408f029dc745f465f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
