+++ Fantasy-Quest/Assets/Scripts/DI/Project/Services/BubbleEventSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: df365df1653434649898e9bfeb526747
+guid: 072ff8cb02ea2e34abafb134e3485385
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
