--- Fantasy-Quest/Assets/Scripts/DialogueBubble/EventSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bc8390d5e6dbc194881f15dc53ddff39
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
