+++ Fantasy-Quest/Assets/Scripts/DialogueBubble/IShowBubble.cs
@@ -1,3 +1,5 @@
+using DI.Project.Services;
+
 namespace DialogueBubble
 {
     internal interface IShowBubble
